+++ R/getParameterSet.R
@@ -53,16 +53,16 @@ getParameterSet.Beta <- function(x, shape1, shape2, verbose = FALSE){
 }
 
 getParameterSet.BetaNoncentral <- function(x, shape1, shape2, location, verbose = FALSE){
-  
+
   if(verbose) message("Parameterised with shape1, shape2 and location.")
-  
-  
+
+
   ps <- ParameterSet$new(id = list("shape1","shape2", "location"), value = list(1,1, 0),
                          support = list(PosReals$new(), PosReals$new(), PosReals$new(zero = TRUE)),
                          settable = list(TRUE, TRUE, TRUE),
                          updateFunc = NULL,
                          description = list("Shape Parameter (alpha)","Shape Parameter (beta)", "Non-centrality parameter"))
-  
+
   return(ps)
 }
 
@@ -132,15 +132,15 @@ getParameterSet.ChiSquared <- function(x, df, verbose = FALSE){
 }
 
 getParameterSet.ChiSquaredNoncentral <- function(x, df, location, verbose = FALSE){
-  
+
   if(verbose) message("Parameterised with df and location.")
-  
+
   ps <- ParameterSet$new(id = list("df", "location"), value = list(1, 0),
                          support = list(PosReals$new(zero = TRUE), PosReals$new(zero = TRUE)),
                          settable = list(TRUE, TRUE),
                          updateFunc = NULL,
                          description = list("Degrees of Freedom", "Non-centrality parameter"))
-  
+
   return(ps)
 }
 
@@ -230,7 +230,7 @@ getParameterSet.FDistribution <- function(x, df1, df2, verbose = FALSE){
 
 getParameterSet.FDistributionNoncentral <- function(x, df1, df2, location, verbose = FALSE){
   if (verbose) message("Parameterised with df1, df2 and location.")
-  
+
   ps <- ParameterSet$new(id = list("df1", "df2", "location"), value = list(1, 1, 0),
                          support = list(PosReals$new(), PosReals$new(), PosReals$new(zero = TRUE)),
                          settable = list(TRUE, TRUE, TRUE),
@@ -716,15 +716,15 @@ getParameterSet.StudentT <- function(x, df, verbose = FALSE){
 }
 
 getParameterSet.StudentTNoncentral <- function(x, df, location, verbose = FALSE){
-  
+
   if(verbose) message("Parameterised with df and location.")
-  
+
   ps <- ParameterSet$new(id = list("df", "location"), value = list(1, 0),
                          support = list(PosReals$new(), Reals$new()),
                          settable = list(TRUE, TRUE),
                          updateFunc = NULL,
                          description = list("Degrees of Freedom", "Non-centrality parameter"))
-  
+
   return(ps)
 }
 
@@ -786,15 +786,24 @@ getParameterSet.Wald <- function(x, mean, shape, verbose = FALSE){
   return(ps)
 }
 
-getParameterSet.Weibull <- function(x, shape, scale, verbose = FALSE){
+getParameterSet.Weibull <- function(x, shape, scale, altscale, verbose = FALSE){
 
-  if(verbose) message("Parameterised with shape and scale.")
+  scale.bool = altscale.bool = FALSE
 
-  ps <-  ParameterSet$new(id = list("shape","scale"), value = list(1,1),
-                          support = list(PosReals$new(), PosReals$new()),
-                          settable = list(TRUE,TRUE),
-                          updateFunc = NULL,
-                          description = list("Shape paramer", "Scale parameter"))
+  if(!is.null(altscale)){
+    if(verbose) message("Parameterised with shape and altscale.")
+    altscale.bool = TRUE
+  } else{
+    if(verbose) message("Parameterised with shape and scale.")
+    scale.bool = TRUE
+  }
+
+  ps <-  ParameterSet$new(id = list("shape","scale","altscale"), value = list(1,1,1),
+                          support = list(PosReals$new(), PosReals$new(), PosReals$new()),
+                          settable = list(TRUE,scale.bool,altscale.bool),
+                          updateFunc = list(NA, NA,
+                                            function(self) self$getParameterValue('scale')^-self$getParameterValue('shape')),
+                          description = list("Shape paramer", "Scale parameter", "Alternate scale parameter"))
 
   return(ps)
 }