+++ NEWS.md
@@ -15,11 +15,15 @@
 
 ### Updated Functions
 
+- Added the common survival parameterisation of Weibull distribution
+- Updated the listX functions to make them quicker and functional when distr6 is not attached
+
 ## Patches
 
 - Added `errormsg` argument to assertions so a custom error message can be provided
 - Fixed error in quantile function in `Empirical` that was causing results to be dropped
 - Fixed bug in `TruncatedDistribution` that prevented multivariate distributions being truncated
+- Fixed error in variance of `Empirical` that was calculating sample not population variance
 
 # distr6 1.1.0